feat(adaptive): on-demand, resizable two-pane (M3 panes)
Rework the list-detail so the list pane is permanent and fills the full width until an item is selected; the detail pane is temporary / on demand - it opens on tap and is dismissed via back (or the detail's own back). No more permanent empty placeholder.

Both panes are weight-flexible and user-resizable via an M3 VerticalDragHandle (drag to change the split; the fraction is remembered across configuration changes), clamped to keep both usable. Extracted into a reusable MediaListDetailScaffold(navController, listPane) so Discover/Feed/Forum can adopt the same pattern. The detail pane still hosts a nested NavHost for correct per-id ViewModel scoping. Refs discussion #64.
